requirements.txt
Django==1.8
Markdown==2.6.2
Pillow==2.9.0
django-appconfig==1.0.1
django-bootstrap-form==3.2
django-compressor==1.5
django-debug-toolbar==1.3.1
six==1.9.0